[K6.0] Wrong layout on choose a template as Default #8559
